Develop

+++ lucaTicket-Eventos/src/main/java/com/ejemplos/spring/repository/EventoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ejemplos.spring.model.Eventos;
 
 @Repository
-public interface EventoRepository extends JpaRepository<Eventos, Long> {
+public interface EventoRepository extends JpaRepository<Eventos, Integer> {
 
 }
